Added new EventProfiler.
