--- Assets/Plugins/CandyCoded/Scripts/CustomEditor/ListEvents.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e6feb84a639b21a43bbec99028ccf8ab
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
